% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Resvidex pipeline}
\usage{
classify(inputFile, model, QC_value = 0.4, Length_value = 0.5, N_value = 2)
}
\arguments{
\item{inputFile}{string path relative to the working directory of the input file. Must be in fasta format.}

\item{model}{A random forest classification model. Must be FULL_GENOME. See \link{FULL_GENOME}.}

\item{QC_value}{numeric value from 0.3 to 1. (default = 0.4)}

\item{Length_value}{numeric value from 0 to 1. Proportion of difference to the expected sequence length. (default = 0.5)}

\item{N_value}{numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)}
}
\value{
a dataframe
}
\description{
Wrapper function that performs all the recommended step. This includes
\describe{
\item{kcounter}{# Count Kmers and normalize data}
\item{prediction_caller}{Run prediction pipeline}
\item{quality_control}{Check Quality of the input data and classification results}
\item{quality_filter}{For sequences with any FLAG present set result Clade to LowQuality}
}
}
\examples{


file_path<-system.file("extdata","test_dataset.fasta",package="resvidex")


Classification<-classify(inputFile=file_path,model=FULL_GENOME)

#the variable Classification has now all the sequences classified
Classification

\dontrun{
#if you want to export the prediction
utils::write.csv2(Classification,"Classification_file.csv")
}
}
