% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_control}
\alias{model_control}
\title{model_control
Utility function to check the classification model object}
\usage{
model_control(model)
}
\arguments{
\item{model}{random forest classification model. Must be FULL_GENOME. See \link{FULL_GENOME}.}
}
\value{
warnings
}
\description{
model_control
Utility function to check the classification model object
}
\examples{
model_control(FULL_GENOME)
}
