% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_caller.R
\name{quality_control}
\alias{quality_control}
\title{quality_control}
\usage{
quality_control(data, QC_value = 0.4, Length_value = 0.5, N_value = 2, model)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=prediction_caller]{prediction_caller()}}}

\item{QC_value}{numeric value from 0.3 to 1. (default = 0.4)}

\item{Length_value}{numeric value from 0 to 1. Proportion of difference to the expected sequence length. (default = 0.5)}

\item{N_value}{numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)}

\item{model}{A random forest classification model. Must be FULL_GENOME. See \link{FULL_GENOME}.}
}
\value{
A list with three logical vectors. In each case TRUE means pass.
}
\description{
quality_control
}
\examples{

file_path<-system.file("extdata","test_dataset.fasta",package="resvidex")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- kcounter(SequenceData=sequence,model=FULL_GENOME)

PredictedData <- prediction_caller(NormalizedData=NormalizedData,model=FULL_GENOME)

quality_control(PredictedData,model=FULL_GENOME)


}
