% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_caller.R
\name{quality_filter}
\alias{quality_filter}
\title{quality_filter}
\usage{
quality_filter(data)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=quality_control]{quality_control()}}}
}
\value{
data.frame
}
\description{
Change Clade definition to "LowQuality" in samples with at least one FLAG from \code{\link[=quality_control]{quality_control()}}.
}
\examples{

file_path<-system.file("extdata","test_dataset.fasta",package="resvidex")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- kcounter(SequenceData=sequence,model=FULL_GENOME)

PredictedData <- prediction_caller(NormalizedData=NormalizedData,model=FULL_GENOME)

PredictedData <- quality_control(PredictedData,model=FULL_GENOME)

PredictedData <- quality_filter(PredictedData)

}
