% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run Resvidex shiny app}
\usage{
run_shiny_app()
}
\value{
a shiny app
}
\description{
Deploys a server that runs the Resvidex app locally
}
\examples{
\dontrun{
resvidex::run_shiny_app()
}

}
