% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runResvidexApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run Resvidex shiny app}
\usage{
runShinyApp()
}
\value{
a shiny app
}
\description{
Deploys a server that runs the Resvidex app locally
}
\examples{
\dontrun{
resvidex::runShinyApp()
}

}
